创建目录替换成工具，增加isdir参数控制

+++ index.js
@@ -7,41 +7,29 @@
 const path = require('path');
 const fs = require('fs');
 const rimraf = require('rimraf');
-let outputPath;
-//创建多层文件夹 同步
-function mkdirsSync(root, dirpath) {
-    if (!fs.existsSync(path.join(root, dirpath))) {
-        var pathtmp = root;
-        dirpath.split(path.sep).forEach(function(dirname) {
-            if (dirname === '.' || dirname === '' || dirname === '..') {
-                return false;
-            }
-            pathtmp = path.join(pathtmp, dirname);
-            if (!fs.existsSync(pathtmp)) {
-                if (!fs.mkdirSync(pathtmp)) {
-                    return false;
-                }
-            }
-        });
-    }
-    return true;
-}
+const mkdirp = require('mkdirp');
+let params = {};
 function WebpackDevServerOutput(options) {
-    outputPath = options.path;
+    params = options;
 }
 WebpackDevServerOutput.prototype.apply = function(compiler) {
   compiler.plugin("emit", function(compilation, callback) {
-      outputPath = outputPath || compiler.output.path;
-      if (fs.existsSync(outputPath)) {
-          rimraf.sync(outputPath);
+      let outputPath = params.path || path.join(path.resolve("."), 'output');
+      if (params.isDel === true) {
+          if (fs.existsSync(outputPath)) {
+              rimraf.sync(outputPath);
+              fs.mkdirSync(outputPath);
+          }
+      } else {
+          if (!fs.existsSync(outputPath)) {
+              fs.mkdirSync(outputPath);
+          }
       }
-      fs.mkdirSync(outputPath);
       for (let filename in compilation.assets) {
           let fileArr = filename.split(path.sep);
-          // 递归创建目录
           if (fileArr.length > 1) {
               let privatePath = fileArr.slice(0, fileArr.length - 1).join(path.sep);
-              mkdirsSync(outputPath, privatePath);
+              mkdirp.sync(path.join(outputPath, privatePath));
           }
           fs.writeFileSync(path.join(outputPath, filename), compilation.assets[filename].source());
       }
